# src/static/swagger-data-analysis-electoral-movements.yaml
swagger: "2.0"
info:
  description: "This API contains election data and movements such as the real time position of candidates, where they are meeting, voting trends, popularity of candidates on social networks and media in order to predict the political actions of other parties and carry out a successful political campaign."
  version: "1.0.0"
  title: "Data analysis and electoral movements"
  contact:
    email: "team2-big-data@data.fr"
host: "127.0.0.1:5000"
basePath: "/v1.0.0"
...
schemes:
- "http"
paths:
...
  /home/daily-vote-tendancy:
    get:
      tags:
      - "tendancy_votes"
      summary: "Returns vote inventories for each candidat in map or in percentage of population"
      description: "Returns a map of status votes trend"
      operationId: "getInventory"
      produces:
      - "application/json"
      parameters: []
      responses:
        "200":
          description: "successfull operation"
          schema:
            type: "object"
            additionalProperties:
              type: "integer"
              format: "int32"
      security:
      - api_key: []
...
securityDefinitions:
  meet_electoral_auth:
    type: "oauth2"
    authorizationUrl: "https://data-analysis-electoral-movements.sourcing.com"
    flow: "implicit"
    scopes:
      write:candidats: "modify our candidat's data in your account"
      read:candidats: "read the other candidats"
  api_key:
    type: "apiKey"
    name: "api_key"
    in: "header"
...